# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "neмытьlE yблюdki",
        "title": "Personal Best API",
        "contact": {
            "name": "raitonoberu",
            "url": "http://raitonobe.ru",
            "email": "raitonoberu@mail.ru"
        },
        "version": "0.1"
    },
    "paths": {
        "/api/competitions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List all competitions from new to old\nFor now there is no way to get start/end but im working on it :)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "competition"
                ],
                "summary": "List competition",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ListCompetitionsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create new competition.\nDays must be different (no same day twice).\nTime must be in format HH:MM.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "competition"
                ],
                "summary": "Create competition",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateCompetitionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.CreateCompetitionResponse"
                        }
                    }
                }
            }
        },
...
        "/api/competitions/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return competition by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "competition"
                ],
                "summary": "Get competition",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetCompetitionResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "competition"
                ],
                "summary": "Delete competition",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of competition",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "competition"
                ],
                "summary": "Update competition",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "comp id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateCompetitionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/competitions/{id}/scores": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get final scores for all competition players",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "competition"
                ],
                "summary": "Get competition scores",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "competition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CompetitionScore"
                            }
                        }
                    }
                }
            }
        },
        "/api/documents": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Save player document. Max file size is 10MB\nFile can be one of:\njpeg, png, webp, gif, pdf",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document"
                ],
                "summary": "Save document",
                "parameters": [
                    {
                        "type": "file",
                        "description": "document",
                        "name": "document",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/documents/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get URL of document",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document"
                ],
                "summary": "Get document",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of document",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetDocumentResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "What do you think it does?",
                "tags": [
                    "document"
                ],
                "summary": "Delete document",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of document",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/users/{user_id}/documents": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List documents of player",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document"
                ],
                "summary": "List player documents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Document"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CompetitionDay": {
            "type": "object",
            "required": [
                "date",
                "end_time",
                "start_time"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "model.CompetitionScore": {
            "type": "object",
            "properties": {
                "lose_score": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/model.GetUserResponse"
                },
                "win_score": {
                    "type": "integer"
                }
            }
        },
        "model.CreateCompetitionRequest": {
            "type": "object",
            "required": [
                "age",
                "closes_at",
                "days",
                "description",
                "name",
                "size",
                "tours"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "closes_at": {
                    "type": "string"
                },
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CompetitionDay"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "tours": {
                    "type": "integer"
                }
            }
        },
        "model.CreateCompetitionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.Document": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.GetCompetitionResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "closes_at": {
                    "type": "string"
                },
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CompetitionDay"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "tours": {
                    "type": "integer"
                },
                "trainer": {
                    "$ref": "#/definitions/model.GetUserResponse"
                }
            }
        },
        "model.GetDocumentResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "model.GetUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/model.PlayerResponse"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
        "model.ListCompetitionsResponse": {
            "type": "object",
            "properties": {
                "competitions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GetCompetitionResponse"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.PlayerResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "is_male": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "preparation": {
                    "type": "string"
                },
                "telegram": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateCompetitionRequest": {
            "type": "object",
            "properties": {
                "closes_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}